[rb] move common code for local drivers into a module
