+++ rb/lib/selenium/webdriver/common.rb
@@ -18,6 +18,7 @@
 # under the License.
 
 require 'selenium/webdriver/common/error'
+require 'selenium/webdriver/common/local_driver'
 require 'selenium/webdriver/common/platform'
 require 'selenium/webdriver/common/proxy'
 require 'selenium/webdriver/common/log_entry'
