+++ rb/lib/selenium/webdriver/common/driver.rb
@@ -314,24 +314,6 @@ module Selenium
         end
       end
 
-      def process_options(options, capabilities)
-        if options && capabilities
-          msg = "Don't use both :options and :capabilities when initializing #{self.class}, prefer :options"
-          raise ArgumentError, msg
-        end
-
-        options ? options.as_json : deprecate_capabilities(capabilities)
-      end
-
-      def deprecate_capabilities(capabilities)
-        unless is_a?(Remote::Driver)
-          WebDriver.logger.deprecate("The :capabilities parameter for #{self.class}",
-                                     ":options argument with an instance of #{self.class}",
-                                     id: :capabilities)
-        end
-        generate_capabilities(capabilities)
-      end
-
       def generate_capabilities(capabilities)
         Array(capabilities).map { |cap|
           if cap.is_a? Symbol
