+++ rb/lib/selenium/webdriver/ie/driver.rb
@@ -29,28 +29,15 @@ module Selenium
       class Driver < WebDriver::Driver
         EXTENSIONS = [DriverExtensions::HasWebStorage].freeze
 
-        def initialize(capabilities: nil, options: nil, service: nil, url: nil, **opts)
-          raise ArgumentError, "Can't initialize #{self.class} with :url" if url
+        include LocalDriver
 
-          caps = process_options(options, capabilities)
-          url = service_url(service || Service.ie)
+        def initialize(capabilities: nil, options: nil, service: nil, url: nil, **opts)
+          caps, url = initialize_local_driver(capabilities, options, service, url)
           super(caps: caps, url: url, **opts)
         end
 
         def browser
           :internet_explorer
-        end
-
-        private
-
-        def process_options(options, capabilities)
-          if options && !options.is_a?(Options)
-            raise ArgumentError, ":options must be an instance of #{Options}"
-          elsif options.nil? && capabilities.nil?
-            options = Options.new
-          end
-
-          super(options, capabilities)
         end
       end # Driver
     end # IE
