+++ rb/lib/selenium/webdriver/safari/driver.rb
@@ -30,28 +30,15 @@ module Selenium
                       DriverExtensions::HasApplePermissions,
                       DriverExtensions::HasWebStorage].freeze
 
-        def initialize(capabilities: nil, options: nil, service: nil, url: nil, **opts)
-          raise ArgumentError, "Can't initialize #{self.class} with :url" if url
+        include LocalDriver
 
-          caps = process_options(options, capabilities)
-          url = service_url(service || Service.safari)
+        def initialize(capabilities: nil, options: nil, service: nil, url: nil, **opts)
+          caps, url = initialize_local_driver(capabilities, options, service, url)
           super(caps: caps, url: url, **opts)
         end
 
         def browser
           :safari
-        end
-
-        private
-
-        def process_options(options, capabilities)
-          if options && !options.is_a?(Options)
-            raise ArgumentError, ":options must be an instance of #{Options}"
-          elsif options.nil? && capabilities.nil?
-            options = Options.new
-          end
-
-          super(options, capabilities)
         end
       end # Driver
     end # Safari
